+++ HISTORY.md
@@ -1,3 +1,10 @@
+# 1.1.3 Settings to Vuex pull on mount
+
+- theme 🌓
+- image Type
+- quote locations (todo) ⏱️
+- quote types (todo)
+
 # 1.1.2 Create the Settings page
 
 -Cat as service calls for image [#30](https://github.com/jimmerydad/cat-quotes/issues/30)
@@ -5,8 +12,7 @@
 - FEAT#4 Quote Type 1: Dad quote initial pull
   [#21](https://github.com/jimmerydad/cat-quotes/issues/21)
 
-- FEAT #2 Settings Page
-- [#18](https://github.com/jimmerydad/cat-quotes/issues/18)
+- FEAT #2 Settings Page [#18](https://github.com/jimmerydad/cat-quotes/issues/18)
 
 # 1.0.2 First apk test